change init

+++ init.hpp
@@ -8,7 +8,7 @@
 #define si(x) ((int)x.size())
 using std::cout, std::cin, std::endl,std::max,std::min,std::swap;
 template<typename T> using vec=std::vector<T>;
-using P = std::pair<i64,int>;
+using P = std::pair<int,int>;
 using i64 = long long;
 
 void solve() {
